requirements.txt
torch>=1.8.0
torchvision>=0.9.0
timm>=0.4.0
opencv-python>=4.5.0
scikit-image>=0.18.0
numpy>=1.19.0
pillow>=8.0.0

